requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.7
django-bootstrap-v5==1.0.11
django-tinymce==3.6.0
Pillow==9.4.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
